move query logic into separate file
